ETL pipeline/requirements.txt
python3-dotenv
pandas
matplotlib
